split cli to subparts
