+++ setup.cfg
@@ -38,11 +38,11 @@ powerpwn =
 
 python_requires = >=3.6,<=3.8.10
 
-scripts = src/powerpwn/cli.py
+scripts = src/powerpwn/main.py
 
 [options.entry_points]
 console_scripts = 
-    powerpwn = powerpwn.cli:main
+    powerpwn = powerpwn.main:main
 
 [mypy]
 plugins = pydantic.mypy
